requirements.txt
polars==0.20.3
pandas==2.1.4
scikit-learn==1.3.2
plotly==5.18.0
kaleido==0.2.1
weasyprint==60.2
openai==1.6.1
watchdog==3.0.0
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0
